feat(TTB-045): Moved functions into gameUtils.js
